Image-Encryption-and-Authentication/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cycler==0.10.0
Django==3.1.2
django-bootstrap4==2.3.1
ImageHash==4.1.0
importlib-metadata==2.0.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
pyparsing==2.4.7
PySimpleGUI==4.30.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scipy==1.5.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
zipp==3.4.0
